Fix multi-channel support with Jack
When Jack is using multiple inputs, the minimum minimal of stereo
outputs is two, as at least two FX channels will be generated.

Also, when the number of audio channels is odd, fluidsynth
(tested with version 2.1.1 and 2.1.5) ends having some problem
setting the connections and causing the Jack thread to terminate.

So, ensure that the number of channels will be an even number.

While here, disable setting the number of channels when Jack
is set to use just one stereo output channel.