Change the UI language translation
Closes #57

This patch adds a combo box at the Options dialog allowing the user to
select and change (after restarting the program) the user interface
language translation, listing the currently installed translations. This
option is persistent between restarts, stored in the user's settings.

There are also some changes about the way the translations path is
calculated depending on the operating system, and the load() method
signature of QTranslator:
https://www.kdab.com/fixing-a-common-antipattern-when-loading-translations-in-qt/

There is only one new translatable string added by this patch. The
language names are displayed in their respective native names, as
provided by QLocale.